update Re layering and simplifying code
