+++ lib/picture_block/interaction/block_sequence.dart
@@ -1,5 +1,5 @@
 // block_sequence.dart
-import 'package:flutter_application_1/picture_block/block_data.dart';
+import 'package:flutter_application_1/picture_block/models/block_data.dart';
 
 class BlockSequence {
   // List to hold the order of blocks
