+++ lib/picture_block/models/block_with_image.dart
@@ -1,0 +1,9 @@
+// lib/picture_block/models/block_with_image.dart
+import 'package:flutter_application_1/picture_block/models/block_data.dart';
+
+class BlockWithImage {
+  final String imagePath;
+  final Shape shape;
+
+  BlockWithImage({required this.imagePath, required this.shape});
+}
