+++ lib/picture_block/models/virtual_tile.dart
@@ -1,0 +1,7 @@
+
+class VirtualTile {
+  final int index;
+  final String tileType;
+
+  VirtualTile(this.index, this.tileType);
+}
