+++ lib/picture_block/pictureblock.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/picture_block/block_sequence.dart';
-import 'package:flutter_application_1/picture_block/virtual_controller.dart';
+import 'package:flutter_application_1/picture_block/interaction/block_sequence.dart';
+import 'package:flutter_application_1/picture_block/interaction/virtual_controller.dart';
 import 'package:provider/provider.dart';
-import 'sidebar.dart'; // Import the Sidebar
-import 'block_data.dart';
-import 'dragable_block.dart';
-import 'block_helpers.dart';
-import 'command_manager.dart';
-import 'category_buttons.dart';
-import 'action_buttons.dart';
+import 'widgets/sidebar.dart'; // Import the Sidebar
+import 'models/block_data.dart';
+import 'widgets/dragable_block.dart';
+import 'interaction/block_helpers.dart';
+import 'widgets/command_manager.dart';
+import 'widgets/category_buttons.dart';
+import 'widgets/action_buttons.dart';
+import 'package:flutter_application_1/picture_block/models/block_with_image.dart';
 
 class PictureBlockPage extends StatefulWidget {
   const PictureBlockPage({Key? key}) : super(key: key);
