+++ lib/picture_block/widgets/dragable_block.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/picture_block/block_shape.dart';
-import 'block_data.dart';
-import 'package:flutter_application_1/picture_block/virtual_controller.dart';
+import 'package:flutter_application_1/picture_block/models/block_shape.dart';
+import '../models/block_data.dart';
+import 'package:flutter_application_1/picture_block/interaction/virtual_controller.dart';
 
 class DraggableBlock extends StatefulWidget {
   final BlockData blockData; // Data associated with the block
