+++ lib/picture_block/widgets/sidebar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/picture_block/virtual_controller.dart';
+import 'package:flutter_application_1/picture_block/interaction/virtual_controller.dart';
 import 'package:provider/provider.dart';
 
 class Sidebar extends StatelessWidget {
